Fix invalid .spv->.bc translation of Intel FPGA merge attribute
This patch introduces fix for the problem related to translation of SPIR-V
decoration with two (or more) Literal Strings to LLVM IR. The processing of
getting Literal Strings from the related decoration was limited to the first
Literal String (first '\0' symbol).
In particular, the merge attribute has two Literal String according to the
specification.

Signed-off-by: Viktoria Maksimova <viktoria.maksimova@intel.com>